More test coverage and refactoring around handlers
